Adds toolbar code into original code as a mobile-only branch
