+++ package.json
@@ -10,6 +10,7 @@
     "@fortawesome/free-solid-svg-icons": "^5.14.0",
     "@fortawesome/react-fontawesome": "^0.1.11",
     "@mui/material": "^5.6.2",
+    "@mui/icons-material": "5.4.2",
     "bootstrap": "^5.1.3",
     "cypress": "^9.5.4",
     "firebase": "^9.6.11",
