+++ src/components/Toolbar/Toolbar.module.scss
@@ -161,21 +161,19 @@ $foodColor: #8dc63f;
     border-left: 1px solid #ccc;
   }
 
-
-
   /*.infoButton {
     top: 17%;
   }*/
 }
 
 .PHLASKButton {
-  transform: scale(3) translateY(-8px);
+  transform: scale(2) translateY(-8px);
 }
 
 .resourceButton {
-  transform: scale(1.2);
+  transform: scale(1);
 }
 
 .contributeButton {
-  transform: scale(1.2);
+  transform: scale(1);
 }
